% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_map_state}
\alias{fars_map_state}
\title{map the distribution of a state in a year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The state number in 1 to 56, representing each state}

\item{year}{The numeric representing the year}
}
\value{
This function will return a map describing the distribution of the accident
}
\description{
This is a function that you can choose a state by the state number and a year, it will return a plot that describe
the distribution of the accident in the state of the choosen year.
}
\examples{
\dontrun{fars_map_state(45,2013)}

}
