% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS Data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the file of the FARS data.}
}
\value{
This funciton returns a tibble containting the FARS data. If an
  incorrect filename is entered the function will stop.
}
\description{
\code{fars_read} reads FARS into the enviroment
}
\details{
This Function reads data from teh US National Highway Trafffice Safety
Administrations' Fatality Analysis reporting System (FARS), give a
filename for the data. It returns a tibble of the data. For this function
to work properly, a filename pointing to an existing file must be give.
}
