% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read_years}
\alias{fars_read_years}
\title{remain only variable month and year of the choosen dataset}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector that contains the year numbers}
}
\value{
The function will return new tibbles with 2 columns(monthe and year), or a warning message:invaild year: \code{year}
and "NULL"
}
\description{
This is a simple function that remain 2 variables of the accident data file of the year you choose, month and year.
When it meets an error, it will return "invaild year: \code{year}"
}
\examples{
\dontrun{fars_read_years(c(2013,2014,2015))}

}
