% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize the count of accidents of each year, each month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector that you can put you numeric years here}
}
\value{
This function will return a summarized table that contains the year you inputed and the accident number of each month
in this year.
}
\description{
This is a fuction that you can enter the year, it will tell you in this year, how many accidents occured in each month
}
\examples{
\dontrun{fars_summarize_years(c(2013,2014))}

}
