% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{make_filename}
\alias{make_filename}
\title{use year to name the file}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A number which is a year}
}
\value{
This function returns the filename as "accident_year.csv.bz2"
}
\description{
This is a simple function that you can input the year and it will return the file name depended on the year
}
\examples{
\dontrun{make_filename(2013)}

}
