% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots selected state number and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{number of state integer or string (coerses to integer)}

\item{year}{integer}
}
\value{
a plot with selected criteria
}
\description{
Takes state number and year and plots a map. Based on the existence of state
number, could throw message "invalid STATE number: state.num". Also if there
are no accidents, shows the message of "no accidents to plot". Uses
previously defined make_filename and fars_read functions.
}
\examples{
fars_map_state(2, 2015)

}
