% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numeric value for the state of the United States to
analyze.}

\item{year}{A numeric value giving the year to analyze.}
}
\value{
This function returns a geographic plot of the state where there has
been at least one fatality traffic accident for a particular year.
}
\description{
This function makes a plot of the fatality traffic for a specified state of
the United States, according to the year established by the user.
}
\note{
There will be errors if any of the arguments are absent or if there are
no numeric values for the arguments.
}
\examples{
fars_map_state(39, 2014)

}
