% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numeric representation of a state}

\item{year}{a character string or numeric representation of a year}
}
\value{
plots a map object of all accidents for the given state in the specified year
}
\description{
given a state and a year,
plot all accident locations for that year in that state
}
\details{
an error will be thrown if an invalid STATE number is supplied.
}
\examples{
\dontrun{
fars_map_state(-1, 1000000)
fars_map_state(1, 2015)
}

}
