% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read}
\alias{fars_read}
\title{read in a data file as a tibble named "data"}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the data file's physical name}
}
\value{
this function returns a tibble data with column number and row number and a dataset named "data", it also
returns "file \code{filename} does not exist".
}
\description{
this is a simple function that firstly find the datafile's name in the current content, if there is not such the file,
the function would return "file dose not exist", if there is the file, the function would read in the data and make it
a tibble table.
}
\examples{
\dontrun{fars_read("accident_2013.csv.bz2")}

}
