% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read the data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{a character string. if the file is in the current working
directory, you do not have to include the path and only the file name with
.csv ending would be enough.}
}
\value{
the data in tbl_df format
}
\description{
This is a function that checks if the file exists and takes couple of
decisions based on. If the file does not exist, it throws a warning message.
If the file exists, it basically imports it as csv and than transforms it
into tbl data frame.
}
