% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Fars package code}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string with the name of the file to read, see
notes.}
}
\value{
A data frame with data readed from the csv file, or an error if the
  file does not exists.
}
\description{
This supporting function reads a csv file.
if the input file doesn't exist execution will stop and
error will be thrown.
}
\note{
To generate file name use: \code{\link{make_filename}}
}
\examples{
\dontrun{
library(dplyr)
library(readr)
yr <- 2015
data <- yr \%>\%
  make_filename \%>\%
  fars_read
View(data)
}
}
\seealso{
\link{make_filename}
}
\author{
Federico Sensi and Coursera
}
