% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the filename the function will
visualize.}
}
\value{
This function returns a data frame of class tibble with the data
from the specified filename. If the filename does not exist, there will
be a message informing about it.
}
\description{
This function shows a quick view of your data in an object of class tibble,
according to a specified CSV filename.
}
\note{
There will be errors if the filename is absent, there are no
quotation marks in the filename argument or the file does not exist.
}
\examples{
fars_read("accident_2013.csv.bz2")

}
