% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{a character string filename mapping to a csv file in the working directory}
}
\value{
a tibble
}
\description{
reads a csv file to a tibble if given filename corresponding to csv file
}
\details{
if character string does not match a filename, error is thrown
}
\examples{
\dontrun{
fars_read('nonsense_name_that_will_throw_error')
fars_read('accident_2015.csv.bz2')
}

}
