% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read years and months from the given file}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{either scalar or vector}
}
\value{
MONTH and year from the accident data based on the availability of
   provided year as a parameter.
}
\description{
This is a function reads accident csv files based on given year vector, and
   then return months and years. This function uses 2 above functions.
}
