% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A numeric vector giving the year or years from which to select.}
}
\value{
This function returns a list with data frames of class tibble with
the months and years from each record of the file.
}
\description{
This function selects the month and year from each record in the files
associated to the US National Highway Traffic Safety Administration's
Fatality Analysis Reporting System, given one or more years.
}
\note{
There will be errors if the year is absent or the if the year is
referenced to a non-existing object in the environment. There are warnings
when the year is invalid (no existing filename with that year).
}
\examples{
fars_read_years(c(2013,2014))

}
