% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector of character or numeric years}
}
\value{
a list of tibbles if all years are valid.
a list of tibbles and NULL values if some years are valid and other are not
a list of NULLs if no years are valid
}
\description{
given a vector of years,
return a list of tibbles corresponding to year with
month and year columns subset
}
\details{
if a year is not found in the data,
a warning message is thrown for that year only
}
\examples{
\dontrun{
fars_read_years(c(2015, 2013,  2014))
fars_read_years(c(1970, 100000,200000))
}

}
