% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize given accident data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{either scalar or vector}
}
\value{
a data.frame of summarized data that is converted to wider format
}
\description{
This is a function that summarize accident csv files based on given year
vector. This function uses fars_read_years function.
}
