% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A numeric vector giving the year or years from which to select}
}
\value{
This function returns a data frame with the amount of times every
month appears per year.
}
\description{
This function summarizes the number of times the months appear for a
specified vector of years.
}
\note{
There will be errors if the year is absent or the if the year is
referenced to a non-existing object in the environment. There are warnings
when the year or years are invalid (no existing filename with that year).
}
\examples{
fars_summarize_years(c(2013,2014,2015))

}
