% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector of character or numeric years}
}
\value{
a tibble
}
\description{
for the years supplied,
return a tibble of the numbers of per-month accident occurrences for those years
}
\details{
an error will be thrown if all supplied years are invalid.
}
\examples{
\dontrun{
fars_summarize_years(c(2015, 2013,  2014))
fars_summarize_years(c(1970, 100000,200000))
}

}
