% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{integer or string in YYYY format}
}
\value{
This function returns a string with provided year input e.g.
    accident_2013.csv.bz2
}
\description{
This function creates a string, which is going to be used for a file name
based on the provided year parameter.
}
\examples{
make_filename(2015)

}
