% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numeric value giving the year to make the filename.}
}
\value{
This function returns the created filename with quotation marks.
}
\description{
This simple function makes a filename for the data associated to the US
National Highway Traffic Safety Administration's Fatality Analysis
Reporting System, according to a specified year.
}
\note{
There will be errors if the year argument is absent or if the year is
referenced to a non-existing object in the environment. If there is a
character string in the year argument, R will introduce NAs by coercion.
}
\examples{
make_filename(2015)

}
