% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a character string or numeric representation of a year}
}
\value{
a character string
}
\description{
makes a filename from a character year input
}
\details{
no errors will be thrown, you can toss this baby anything ;)
}
\examples{
\dontrun{
make_filename('absolute_nonsens')
make_filename('1970')
}

}
